pip-requirements.txt
bcrypt==4.1.3
cffi==1.16.0
cryptography==42.0.8
paramiko==3.4.0
pycparser==2.22
pynacl==1.5.0
packaging==24.1
pytz==2024.1
python-dateutil==2.9.0
sshtunnel==0.4.0
prettytable==3.10.0
wcwidth==0.2.13
fsspec==2024.6.1
torchdata==0.7.1
tzdata==2024.1
pandas==2.2.2
annotated-types==0.7.0
pydantic==2.8.2
pydantic-core==2.20.1
joblib==1.4.2
littleutils==0.2.2
ogb==1.3.6
outdated==0.2.2
scikit-learn==1.5.1
threadpoolctl==3.5.0
dnspython==2.6.1
pymongo==4.8.0